Fixed modular controller. Started modular battery
In order to be able to switch between the different controllers (UNO, Mega, nano, pico etc) I created a modular top that can screw and unscrew a piece of it.
For the battery it is a similar concept, in order to be able to switch between different casings, I just created an individual module.